Allow registering a Func as a handler
The StaticMessageContextHandler registration did not allow passing an anonymous method. The generated method was not static, so it would fail the validation.
We have removed the validation, and the necessity to have a Type specified.
Since the Type is now optional (for a valid subscription), there is a new property called HandlerName to use for instrumentation.
For instance method handlers, it is the name of the type.
For method group static methods, you must now specify a description, to disambiguate from the overload that takes an Expression.
The new overload that takes an Expression allows specifying a lambda that invokes a method that does not match the StaticMessageContextHandler signature. The description is inferred based on the method invoked.